DAPA-54 | fix: HomepageView Carousel
